backend/fixieAuth/requirements.txt
djangorestframework
djangorestframework-simplejwt
gunicorn==23.0.0
cryptography
mssql-django
pyodbc
hypothesis
pytest-django